Extract LocalizedTemplateContentReaderFactory
