+++ framework/src/Volo.Abp.TextTemplating/Volo/Abp/TextTemplating/VirtualFiles/ILocalizedTemplateContentReaderFactory.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Volo.Abp.TextTemplating.VirtualFiles
+{
+    public interface ILocalizedTemplateContentReaderFactory
+    {
+        Task<ILocalizedTemplateContentReader> CreateAsync(TemplateDefinition templateDefinition);
+    }
+}
